+++ Changelog.md
@@ -1,5 +1,7 @@
 # Version 1.3.9
 - Improvements for when a separate throttle is controlling the same loco
+- move to 'newer' function command
+- Bug fix for functions. Direction would toggle would a function button was pressed
 
 # Version 1.3.8
 - Addition of a CV Programmer page